Typhon
Typhon is a monstrous figure in Greek mythology, often described as one of the deadliest creatures in ancient tales. According to myth, he was the offspring of Gaia (Earth) and Tartarus (the deep abyss), embodying the primordial chaos and destruction. Typhon was depicted as a gigantic and fearsome entity with a hundred dragon heads, each capable of spewing fire and venom. His lower body consisted of massive serpentine coils, making him an imposing and terrifying sight. Typhon's roar was said to cause earthquakes and his breath to incite storms, representing nature's untamed and catastrophic forces.   https://typhonmachinery.com/